WS-NA: Moves files from components to lib and renames [copilot]
