+++ src/app/lib/analyticsUtils/beaconModels/index.client.test.ts
@@ -1,7 +1,7 @@
 import { ResonanceMode } from '@bbc/resonance';
 import { Platforms } from '#app/models/types/global';
 import * as getEnvConfigModule from '#app/lib/utilities/getEnvConfig';
-import * as genericLabelHelpers from '../../../lib/analyticsUtils';
+import * as genericLabelHelpers from '..';
 import {
   buildResonanceAnalyticsModel,
   buildActivationEventModel,
@@ -22,7 +22,7 @@ const analyticsUtilFunctions = [
   { name: 'sanitise', source: genericLabelHelpers },
 ];
 
-describe('atiUrl', () => {
+describe('beaconModels', () => {
   beforeEach(() => {
     analyticsUtilFunctions.forEach(func => {
       mockAndSet(func, func.name);
